Add files via upload
Unable to record a video, chrome is up to date but recordscreen.io isn't cooperating. 

Please introduce yourself and the topic you choose do your website on. - Alex Danielson - Coffee Bar Search Site
Who is the target audience? - Coffee consumers, baristas, roasters, and shop owners.
What are the main goals/objectives of your site and how did the design of your site meet the objectives? - My site is in designed to create a web of recommended shops for consumers to check out based on location and suggestions from shop owners and people of various positions in industry. Currently the site does not have the functionality needed to fulfill this task.
Briefly describe you overall typography and color scheme and the decisions that motivated your choose of color, typography. Typography is standard default Verdana, the color scheme is reflective of what I thought worked best with images used for a particular screen.
Please navigate throughout each page of your site describing the content and layout. N/A
What three or more areas did you enhance or change from your assignment  4? N/A
If implemented, demonstrate your site’s responsiveness. N/A
Please explain areas of difficulty as it related to finishing your site - Aside from css in the beginning the semester itself was difficult albeit fun. namely with this class.
Do you have plans for the site after the course is over? I certainly plan to keep working on this despite me considering withdrawal around midterms I did have a good time learning this and would like to see what it can do. There is an attempt at media query on the home page for an image change at breakpoint, and the beginnings of a slide show thanks to w3schools under the suggestions page.